refactored utils as es6 modules
