+++ src/drawer.canvas.js
@@ -1,4 +1,5 @@
 'use strict';
+import * as util from './util';
 
 WaveSurfer.Drawer.Canvas = Object.create(WaveSurfer.Drawer);
 
@@ -98,7 +99,7 @@ WaveSurfer.util.extend(WaveSurfer.Drawer.Canvas, {
 
         var absmax = 1;
         if (this.params.normalize) {
-            absmax = WaveSurfer.util.max(peaks);
+            absmax = util.max(peaks);
         }
 
         var scale = length / width;
@@ -156,8 +157,8 @@ WaveSurfer.util.extend(WaveSurfer.Drawer.Canvas, {
 
         var absmax = 1;
         if (this.params.normalize) {
-            var max = WaveSurfer.util.max(peaks);
-            var min = WaveSurfer.util.min(peaks);
+            var max = util.max(peaks);
+            var min = util.min(peaks);
             absmax = -min > max ? -min : max;
         }
