+++ src/mediaelement.js
@@ -1,4 +1,5 @@
 'use strict';
+import * as util from './util';
 
 WaveSurfer.MediaElement = Object.create(WaveSurfer.WebAudio);
