+++ src/util/index.js
@@ -1,0 +1,6 @@
+export { default as ajax } from './ajax';
+export { default as getId } from './get-id';
+export { default as max } from './max';
+export { default as min } from './min';
+export { default as observer } from './observer';
+export { default as extend } from './extend';
